Extract abstract superclass
This is a nicer pattern than inheritance from a concrete class.
