+++ enterprise/server-enterprise/src/test/java/org/neo4j/server/rest/security/RESTInteraction.java
@@ -20,341 +20,65 @@
 package org.neo4j.server.rest.security;
 
 import org.codehaus.jackson.JsonNode;
-import org.codehaus.jackson.node.ArrayNode;
-import org.codehaus.jackson.node.IntNode;
-import org.codehaus.jackson.node.LongNode;
-import org.codehaus.jackson.node.ObjectNode;
-import org.codehaus.jackson.node.TextNode;
 
 import java.io.IOException;
-import java.util.ArrayList;
-import java.util.Iterator;
 import java.util.Map;
-import java.util.TreeMap;
 import java.util.function.Consumer;
 import javax.ws.rs.core.HttpHeaders;
 
-import org.neo4j.bolt.BoltKernelExtension;
 import org.neo4j.graphdb.ResourceIterator;
-import org.neo4j.graphdb.factory.GraphDatabaseSettings;
-import org.neo4j.io.fs.DefaultFileSystemAbstraction;
-import org.neo4j.io.fs.FileSystemAbstraction;
-import org.neo4j.kernel.api.KernelTransaction;
-import org.neo4j.kernel.api.security.AuthSubject;
-import org.neo4j.kernel.impl.coreapi.InternalTransaction;
-import org.neo4j.kernel.impl.factory.GraphDatabaseFacade;
-import org.neo4j.server.enterprise.helpers.EnterpriseServerBuilder;
-import org.neo4j.server.helpers.CommunityServerBuilder;
-import org.neo4j.server.rest.domain.JsonHelper;
-import org.neo4j.server.rest.domain.JsonParseException;
-import org.neo4j.kernel.enterprise.api.security.EnterpriseAuthManager;
-import org.neo4j.server.security.enterprise.auth.EnterpriseAuthAndUserManager;
-import org.neo4j.server.security.enterprise.auth.EnterpriseUserManager;
-import org.neo4j.server.security.enterprise.auth.NeoInteractionLevel;
 import org.neo4j.test.server.HTTP;
 
-import static org.hamcrest.Matchers.containsString;
-import static org.hamcrest.Matchers.equalTo;
-import static org.hamcrest.Matchers.not;
-import static org.junit.Assert.assertThat;
-import static org.junit.Assert.fail;
-import static org.neo4j.graphdb.factory.GraphDatabaseSettings.BoltConnector.EncryptionLevel.OPTIONAL;
-import static org.neo4j.graphdb.factory.GraphDatabaseSettings.boltConnector;
-import static org.neo4j.kernel.api.security.AuthToken.newBasicAuthToken;
+import static org.neo4j.test.server.HTTP.RawPayload.quotedJson;
 
-class RESTInteraction extends CommunityServerTestBase implements NeoInteractionLevel<RESTSubject>
+class RESTInteraction extends AbstractRESTInteraction
 {
+
+    RESTInteraction( Map<String,String> config ) throws IOException
+    {
+        super( config );
+    }
+
+    @Override
     String commitPath()
     {
         return "db/data/transaction/commit";
     }
 
-    static final String POST = "POST";
-
-    private EnterpriseAuthManager authManager;
-
-    RESTInteraction( Map<String,String> config ) throws IOException
+    @Override
+    HTTP.RawPayload constructQuery( String query )
     {
-        CommunityServerBuilder builder = EnterpriseServerBuilder.server();
-        builder = builder
-                .withProperty( boltConnector( "0" ).type.name(), "BOLT" )
-                .withProperty( boltConnector( "0" ).enabled.name(), "true" )
-                .withProperty( boltConnector( "0" ).encryption_level.name(), OPTIONAL.name() )
-                .withProperty( BoltKernelExtension.Settings.tls_key_file.name(),
-                        NeoInteractionLevel.tempPath( "key", ".key" ) )
-                .withProperty( BoltKernelExtension.Settings.tls_certificate_file.name(),
-                        NeoInteractionLevel.tempPath( "cert", ".cert" ) )
-                .withProperty( GraphDatabaseSettings.auth_enabled.name(), Boolean.toString( true ) );
-
-        for ( Map.Entry<String,String> entry : config.entrySet() )
-        {
-            builder = builder.withProperty( entry.getKey(), entry.getValue() );
-        }
-        this.server = builder.build();
-        this.server.start();
-        authManager = this.server.getDependencyResolver().resolveDependency( EnterpriseAuthManager.class );
+        return quotedJson( "{'statements':[{'statement':'" +
+                           query.replace( "'", "\\'" ).replace( "\"", "\\\"" )
+                           + "'}]}" );
     }
 
     @Override
-    public EnterpriseUserManager getLocalUserManager() throws Exception
-    {
-        if ( authManager instanceof EnterpriseAuthAndUserManager )
-        {
-            return ((EnterpriseAuthAndUserManager) authManager).getUserManager();
-        }
-        throw new Exception("The used configuration does not have a user manager");
-    }
-
-    @Override
-    public GraphDatabaseFacade getLocalGraph()
-    {
-        return server.getDatabase().getGraph();
-    }
-
-    @Override
-    public FileSystemAbstraction fileSystem()
-    {
-        return new DefaultFileSystemAbstraction();
-    }
-
-    @Override
-    public InternalTransaction beginLocalTransactionAsUser( RESTSubject subject, KernelTransaction.Type txType )
-            throws Throwable
-    {
-        AuthSubject authSubject = authManager.login( newBasicAuthToken( subject.username, subject.password ) );
-        return getLocalGraph().beginTransaction( txType, authSubject );
-    }
-
-    @Override
-    public String executeQuery( RESTSubject subject, String call, Map<String,Object> params,
-            Consumer<ResourceIterator<Map<String, Object>>> resultConsumer )
-    {
-        HTTP.RawPayload payload = constructQuery( call );
-        HTTP.Response response = HTTP.withHeaders( HttpHeaders.AUTHORIZATION, subject.principalCredentials )
-                .request( POST, commitURL(), payload );
-
-        try
-        {
-            String error = parseErrorMessage( response );
-            if ( !error.isEmpty() )
-            {
-                return error;
-            }
-            consume( resultConsumer, JsonHelper.jsonNode( response.rawContent() ) );
-        }
-        catch ( JsonParseException e )
-        {
-            fail( "Unexpected error parsing Json!" );
-        }
-
-        return "";
-    }
-
     void consume( Consumer<ResourceIterator<Map<String,Object>>> resultConsumer, JsonNode data )
     {
         if ( data.has( "results" ) && data.get( "results" ).has( 0 ) )
         {
-            JsonNode firstResult = data.get( "results" ).get( 0 );
-            RESTResult result = new RESTResult( firstResult );
-            resultConsumer.accept( result );
+            resultConsumer.accept( new RESTResult( data.get( "results" ).get( 0 ) ) );
         }
     }
 
-    HTTP.RawPayload constructQuery( String query )
-    {
-        return HTTP.RawPayload.quotedJson( "{'statements':[{'statement':'" +
-                                           query.replace( "'", "\\'" ).replace( "\"", "\\\"" )
-                                           + "'}]}" );
-    }
-
     @Override
-    public RESTSubject login( String username, String password ) throws Exception
-    {
-        String principalCredentials = challengeResponse( username, password );
-        return new RESTSubject( username, password, principalCredentials );
-    }
-
     protected HTTP.Response authenticate( String principalCredentials )
     {
         return HTTP.withHeaders( HttpHeaders.AUTHORIZATION, principalCredentials ).request( POST, commitURL() );
     }
 
-    @Override
-    public void logout( RESTSubject subject ) { }
-
-    @Override
-    public void updateAuthToken( RESTSubject subject, String username, String password )
+    private class RESTResult extends AbstractRESTResult
     {
-        subject.principalCredentials = challengeResponse( username, password );
-    }
-
-    @Override
-    public String nameOf( RESTSubject subject )
-    {
-        return subject.username;
-    }
-
-    @Override
-    public void tearDown()
-    {
-        if ( server != null )
-        {
-            server.stop();
-        }
-    }
-
-    @Override
-    public void assertAuthenticated( RESTSubject subject )
-    {
-        assertThat( authenticate( subject.principalCredentials ).status(), equalTo( 200 ) );
-    }
-
-    @Override
-    public void assertPasswordChangeRequired( RESTSubject subject ) throws Exception
-    {
-        HTTP.Response response = authenticate( subject.principalCredentials );
-        assertThat( response.status(), equalTo( 403 ) );
-        assertThat( parseErrorMessage( response ), containsString( "User is required to change their password." ) );
-    }
-
-    @Override
-    public void assertInitFailed( RESTSubject subject )
-    {
-        assertThat( authenticate( subject.principalCredentials ).status(), not( equalTo( 200 ) ) );
-    }
-
-    @Override
-    public void assertSessionKilled( RESTSubject subject )
-    {
-        // There is no session that could have been killed
-    }
-
-    @Override
-    public String getConnectionDetails()
-    {
-        return "server-session";
-    }
-
-    String parseErrorMessage( HTTP.Response response )
-    {
-        try
-        {
-            JsonNode data = JsonHelper.jsonNode( response.rawContent() );
-            if ( data.has( "errors" ) && data.get( "errors" ).has( 0 ) )
-            {
-                JsonNode firstError = data.get( "errors" ).get( 0 );
-                if ( firstError.has( "message" ) )
-                {
-                    return firstError.get( "message" ).asText();
-                }
-            }
-        }
-        catch ( JsonParseException e )
-        {
-            fail( "Unexpected error parsing Json!" );
-        }
-        return "";
-    }
-
-    String commitURL()
-    {
-        return server.baseUri().resolve( commitPath() ).toString();
-    }
-
-    class RESTResult implements ResourceIterator<Map<String,Object>>
-    {
-        private JsonNode data;
-        private JsonNode columns;
-        private int index = 0;
-
         RESTResult( JsonNode fullResult )
         {
-            this.data = fullResult.get( "data" );
-            this.columns = fullResult.get( "columns" );
+            super( fullResult );
         }
 
         @Override
-        public void close()
-        {
-            index = data.size();
-        }
-
-        @Override
-        public boolean hasNext()
-        {
-            return index < data.size();
-        }
-
-        @Override
-        public Map<String,Object> next()
-        {
-            JsonNode row = getRow( data, index++ );
-            TreeMap<String,Object> map = new TreeMap<>();
-            for ( int i = 0; i < columns.size(); i++ )
-            {
-                String key = columns.get( i ).asText();
-                Object value = getValue( row.get( i ) );
-                map.put( key, value );
-            }
-            return map;
-        }
-
-        JsonNode getRow( JsonNode data, int i )
+        protected JsonNode getRow( JsonNode data, int i )
         {
             return data.get( i ).get( "row" );
         }
-    }
-
-    private Object getValue( JsonNode valueNode )
-    {
-        Object value;
-
-        if ( valueNode instanceof TextNode )
-        {
-            value = valueNode.asText();
-        }
-        else if ( valueNode instanceof ObjectNode )
-        {
-            value = mapValue( valueNode.getFieldNames(), valueNode );
-        }
-        else if ( valueNode instanceof ArrayNode )
-        {
-            ArrayNode aNode = (ArrayNode) valueNode;
-            ArrayList<String> listValue = new ArrayList<>( aNode.size() );
-            for ( int j = 0; j < aNode.size(); j++ )
-            {
-                listValue.add( aNode.get( j ).asText() );
-            }
-            value = listValue;
-        }
-        else if ( valueNode instanceof IntNode )
-        {
-            value = valueNode.getIntValue();
-        }
-        else if ( valueNode instanceof LongNode )
-        {
-            value = valueNode.getLongValue();
-        }
-        else
-        {
-            throw new RuntimeException( String.format(
-                "Unhandled REST value type '%s'. Need String (TextNode), List (ArrayNode), Object (ObjectNode), long (LongNode), or int (IntNode).",
-                valueNode.getClass()
-            ) );
-        }
-        return value;
-    }
-
-    private Map<String,Object> mapValue( Iterator<String> columns, JsonNode node )
-    {
-        TreeMap<String,Object> map = new TreeMap<>();
-        while ( columns.hasNext() )
-        {
-            String key = columns.next();
-            Object value = getValue( node.get( key ) );
-            map.put( key, value );
-        }
-
-        return map;
     }
 }
